Feat: Refactor reservoir and river basin components to use new lookup mapping and API structure
