+++ front/src/common/mappers/index.ts
@@ -1,0 +1,1 @@
+export * from "./lookup.mappers";
