+++ front/src/common/models/lookup.model.ts
@@ -2,3 +2,8 @@ export interface Lookup {
   id: string;
   name: string;
 }
+
+export interface LookupApi {
+  _id: string;
+  nombre: string;
+}
