+++ front/src/core/constants/index.ts
@@ -1,1 +1,2 @@
 export * from "./provincias.constants";
+export * from "./cuencas.constants";
