+++ front/src/lib/mongodb.ts
@@ -2,29 +2,32 @@ import { MongoClient, type Db } from "mongodb";
 
 //hack de TypeScript para tipar globalThis con nuestra propiedad custom _mongoClient
 const globalForMongo = globalThis as typeof globalThis & {
-    _mongoClient?: MongoClient;
+  _mongoClient?: MongoClient;
 };
 
 //crea el cliente solo si no existe ya (singleton)
-async function  getClient(): Promise<MongoClient> {
-    const connectionString = process.env.MONGODB_CONNECTION_STRING;
-    if (!connectionString) {
-        throw new Error(
-            "Please define the MONGODB_CONNECTION_STRING environment variable in .env.local"
-        );
-    }
+async function getClient(): Promise<MongoClient> {
+  const connectionString = process.env.MONGODB_CONNECTION_STRING;
+  if (!connectionString) {
+    throw new Error(
+      "Please define the MONGODB_CONNECTION_STRING environment variable in .env.local",
+    );
+  }
 
-    if (!globalForMongo._mongoClient) {
-        console.log("[mongodb] Connecting to MongoDB...", connectionString.replace(/\/\/.*@/, "//***@"));
-        globalForMongo._mongoClient = new MongoClient(connectionString);
-        await globalForMongo._mongoClient.connect();
-        console.log("[mongodb] Connected successfully");
-    }
+  if (!globalForMongo._mongoClient) {
+    console.log(
+      "[mongodb] Connecting to MongoDB...",
+      connectionString.replace(/\/\/.*@/, "//***@"),
+    );
+    globalForMongo._mongoClient = new MongoClient(connectionString);
+    await globalForMongo._mongoClient.connect();
+    console.log("[mongodb] Connected successfully");
+  }
 
-    return globalForMongo._mongoClient;
+  return globalForMongo._mongoClient;
 }
 //lo que exportamo devuelve la instancia de Db lista para hacer queries
 export async function getDb(): Promise<Db> {
-    const client = await getClient();
-    return client.db();
+  const client = await getClient();
+  return client.db();
 }
